use internal buffer for transmission phase
The previous "pub sub" implementation is incompatible with using the
shared internal buffer, so it's been refactored away to more closely
resemble the synchronous request/response flow from the option phase.

Moving over to the internal buffer is necessary but not sufficient for
moving the memory allocations to the edge because the Conn's Read method
doesn't allow for client code to supply a buffer to write the data to.

Because of this, I've decided to break the API to allow callers to
supply a buffer to the Conn's Read method.

Since this change is backwards incompatible, might as well do the
following breaking changes:

* Update the Read method to accept a caller-supplied buffer for data
  and also return the number of bytes written to it so clients can
  detect partial reads.
* Relocate the transmission phase methods' CommandFlags parameter to be
  last (mainly cosmetic)
* Relocate the transmission phase methods' parameters to more closely
  resemble methods from the io package (e.g., byte buffer, then offset.)
* Change the BlockStatus method to be able to return multiple Status
  objects, since the protocol spec mentions that "For a successful
  return, the server MUST use a structured reply, containing exactly one
  chunk of type NBD_REPLY_TYPE_BLOCK_STATUS per selected context id"

Resolves a similar condition as:

	9547ea3 ("introduce internal buffer for message (de)serialization")
	Reported-by: m01e (memory allocation denial-of-service)